+++ src/simputils/config/components/prisms/ObjConfigStorePrism.py
@@ -6,7 +6,8 @@ from simputils.config.enums import ConfigStoreType
 from simputils.config.exceptions import StrictKeysEnabled
 
 
-class ObjConfigStorePrism:
+class ObjConfigStorePrism:  # pragma: no cover
+    # MARK  No Cover might be temporary here
 
     _config_store = None